review K3849 - comment 1 bullet proofing

+++ NbtLogic/csw/Schema/scripts/prod/CswUpdateSchemaCase27779_part2.cs
@@ -17,7 +17,10 @@ namespace ChemSW.Nbt.Schema
         {
 
             //delete the length column
-            _CswNbtSchemaModTrnsctn.dropColumn( "nodetype_props", "length" );
+            if( _CswNbtSchemaModTrnsctn.isColumnDefined( "nodetype_props", "length" ) )
+            {
+                _CswNbtSchemaModTrnsctn.dropColumn( "nodetype_props", "length" );
+            }
 
         }//Update()
